fix broken tests

+++ test/unit/test_data.py
@@ -21,12 +21,12 @@ from tgm.timedelta import TimeDeltaDG
 
 
 def test_init_dg_data():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     data = DGData.from_raw(edge_timestamps, edge_index)
     torch.testing.assert_close(data.edge_index, edge_index)
     torch.testing.assert_close(data.timestamps, edge_timestamps)
-    torch.testing.assert_close(data.edge_event_idx, torch.LongTensor([0, 1]))
+    torch.testing.assert_close(data.edge_event_idx, torch.IntTensor([0, 1]))
     assert data.edge_feats is None
     assert data.node_event_idx is None
     assert data.node_ids is None
@@ -36,28 +36,28 @@ def test_init_dg_data():
 
 
 def test_init_dg_data_with_time_delta():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     data = DGData.from_raw(edge_timestamps, edge_index, time_delta=TimeDeltaDG('s'))
     assert data.time_delta == TimeDeltaDG('s')
 
 
 def test_init_dg_data_with_string_time_delta():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     data = DGData.from_raw(edge_timestamps, edge_index, time_delta='s')
     assert data.time_delta == TimeDeltaDG('s')
 
 
 def test_init_dg_data_no_node_events_with_edge_features():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     edge_feats = torch.rand(2, 5)
     data = DGData.from_raw(edge_timestamps, edge_index, edge_feats)
     torch.testing.assert_close(data.edge_index, edge_index)
     torch.testing.assert_close(data.timestamps, edge_timestamps)
     torch.testing.assert_close(data.edge_feats, edge_feats)
-    torch.testing.assert_close(data.edge_event_idx, torch.LongTensor([0, 1]))
+    torch.testing.assert_close(data.edge_event_idx, torch.IntTensor([0, 1]))
     assert data.node_event_idx is None
     assert data.node_ids is None
     assert data.dynamic_node_feats is None
@@ -66,19 +66,19 @@ def test_init_dg_data_no_node_events_with_edge_features():
 
 
 def test_init_dg_data_node_events():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     edge_feats = torch.rand(2, 5)
-    node_ids = torch.LongTensor([1, 2, 3])
-    node_timestamps = torch.LongTensor([6, 7, 8])
+    node_ids = torch.IntTensor([1, 2, 3])
+    node_timestamps = torch.IntTensor([6, 7, 8])
     data = DGData.from_raw(
         edge_timestamps, edge_index, edge_feats, node_timestamps, node_ids
     )
     torch.testing.assert_close(data.edge_index, edge_index)
-    torch.testing.assert_close(data.timestamps, torch.LongTensor([1, 5, 6, 7, 8]))
+    torch.testing.assert_close(data.timestamps, torch.IntTensor([1, 5, 6, 7, 8]))
     torch.testing.assert_close(data.edge_feats, edge_feats)
-    torch.testing.assert_close(data.edge_event_idx, torch.LongTensor([0, 1]))
-    torch.testing.assert_close(data.node_event_idx, torch.LongTensor([2, 3, 4]))
+    torch.testing.assert_close(data.edge_event_idx, torch.IntTensor([0, 1]))
+    torch.testing.assert_close(data.node_event_idx, torch.IntTensor([2, 3, 4]))
     torch.testing.assert_close(data.node_ids, node_ids)
     assert data.dynamic_node_feats is None
     assert data.static_node_feats is None
@@ -86,11 +86,11 @@ def test_init_dg_data_node_events():
 
 
 def test_init_dg_data_node_events_and_node_features():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     edge_feats = torch.rand(2, 5)
-    node_ids = torch.LongTensor([1, 2, 3])
-    node_timestamps = torch.LongTensor([6, 7, 8])
+    node_ids = torch.IntTensor([1, 2, 3])
+    node_timestamps = torch.IntTensor([6, 7, 8])
     dynamic_node_feats = torch.rand(3, 7)
     static_node_feats = torch.rand(21, 11)
     data = DGData.from_raw(
@@ -103,10 +103,10 @@ def test_init_dg_data_node_events_and_node_features():
         static_node_feats,
     )
     torch.testing.assert_close(data.edge_index, edge_index)
-    torch.testing.assert_close(data.timestamps, torch.LongTensor([1, 5, 6, 7, 8]))
+    torch.testing.assert_close(data.timestamps, torch.IntTensor([1, 5, 6, 7, 8]))
     torch.testing.assert_close(data.edge_feats, edge_feats)
-    torch.testing.assert_close(data.edge_event_idx, torch.LongTensor([0, 1]))
-    torch.testing.assert_close(data.node_event_idx, torch.LongTensor([2, 3, 4]))
+    torch.testing.assert_close(data.edge_event_idx, torch.IntTensor([0, 1]))
+    torch.testing.assert_close(data.node_event_idx, torch.IntTensor([2, 3, 4]))
     torch.testing.assert_close(data.node_ids, node_ids)
     torch.testing.assert_close(data.dynamic_node_feats, dynamic_node_feats)
     torch.testing.assert_close(data.static_node_feats, static_node_feats)
@@ -114,11 +114,11 @@ def test_init_dg_data_node_events_and_node_features():
 
 
 def test_init_dg_data_sort_required():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([5, 1])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([5, 1])
     edge_feats = torch.rand(2, 5)
-    node_ids = torch.LongTensor([1, 2, 3])
-    node_timestamps = torch.LongTensor([8, 7, 6])
+    node_ids = torch.IntTensor([1, 2, 3])
+    node_timestamps = torch.IntTensor([8, 7, 6])
     dynamic_node_feats = torch.rand(3, 7)
     static_node_feats = torch.rand(21, 11)
     data = DGData.from_raw(
@@ -131,8 +131,8 @@ def test_init_dg_data_sort_required():
         static_node_feats,
     )
 
-    exp_edge_index = torch.LongTensor([[10, 20], [2, 3]])
-    exp_node_ids = torch.LongTensor([3, 2, 1])
+    exp_edge_index = torch.IntTensor([[10, 20], [2, 3]])
+    exp_node_ids = torch.IntTensor([3, 2, 1])
     exp_edge_feats = torch.Tensor([edge_feats[1].tolist(), edge_feats[0].tolist()])
     exp_dynamic_node_feats = torch.Tensor(
         [
@@ -142,12 +142,12 @@ def test_init_dg_data_sort_required():
         ]
     )
     torch.testing.assert_close(data.edge_index, exp_edge_index)
-    torch.testing.assert_close(data.timestamps, torch.LongTensor([1, 5, 6, 7, 8]))
+    torch.testing.assert_close(data.timestamps, torch.IntTensor([1, 5, 6, 7, 8]))
     torch.testing.assert_close(data.edge_feats, exp_edge_feats)
-    torch.testing.assert_close(data.edge_event_idx, torch.LongTensor([1, 0]))
+    torch.testing.assert_close(data.edge_event_idx, torch.IntTensor([1, 0]))
     torch.testing.assert_close(
         data.node_event_idx,
-        torch.LongTensor(
+        torch.IntTensor(
             [4, 3, 2],
         ),
     )
@@ -158,15 +158,15 @@ def test_init_dg_data_sort_required():
 
 
 def test_init_dg_data_bad_args_invalid_node_id():
-    edge_index = torch.LongTensor([[PADDED_NODE_ID, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[PADDED_NODE_ID, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     with pytest.raises(InvalidNodeIDError):
         _ = DGData.from_raw(edge_timestamps, edge_index)
 
-    edge_index = torch.LongTensor([[1, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
-    node_ids = torch.LongTensor([PADDED_NODE_ID])
-    node_timestamps = torch.LongTensor([1])
+    edge_index = torch.IntTensor([[1, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
+    node_ids = torch.IntTensor([PADDED_NODE_ID])
+    node_timestamps = torch.IntTensor([1])
     with pytest.raises(InvalidNodeIDError):
         _ = DGData.from_raw(
             edge_timestamps,
@@ -186,27 +186,27 @@ def test_init_dg_data_bad_args_empty_graph():
 
 def test_init_dg_data_bad_args_bad_timestamps():
     # Negative timestamps not supported
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
     with pytest.raises(ValueError):
-        _ = DGData.from_raw(torch.LongTensor([-1, 5]), edge_index)
+        _ = DGData.from_raw(torch.IntTensor([-1, 5]), edge_index)
 
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
     with pytest.raises(TypeError):
         _ = DGData.from_raw('foo', edge_index)
 
 
 def test_init_dg_data_bad_args_bad_edge_index():
-    edge_timestamps = torch.LongTensor([-1, 5])
+    edge_timestamps = torch.IntTensor([-1, 5])
     with pytest.raises(TypeError):
         _ = DGData.from_raw(edge_timestamps, 'foo')
 
     with pytest.raises(ValueError):
-        _ = DGData.from_raw(edge_timestamps, torch.LongTensor([1, 2]))
+        _ = DGData.from_raw(edge_timestamps, torch.IntTensor([1, 2]))
 
 
 def test_init_dg_data_bad_args_bad_edge_feats():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([-1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([-1, 5])
     with pytest.raises(TypeError):
         _ = DGData.from_raw(edge_timestamps, edge_index, 'foo')
 
@@ -215,10 +215,10 @@ def test_init_dg_data_bad_args_bad_edge_feats():
 
 
 def test_init_dg_data_bad_args_bad_node_ids():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     edge_feats = torch.rand(2, 5)
-    node_timestamps = torch.LongTensor([6, 7, 8])
+    node_timestamps = torch.IntTensor([6, 7, 8])
 
     with pytest.raises(TypeError):
         _ = DGData.from_raw(
@@ -231,15 +231,15 @@ def test_init_dg_data_bad_args_bad_node_ids():
             edge_index,
             edge_feats,
             node_timestamps,
-            node_ids=torch.LongTensor([0]),
+            node_ids=torch.IntTensor([0]),
         )
 
 
 def test_init_dg_data_bad_args_non_integral_types():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
-    node_timestamps = torch.LongTensor([6, 7, 8])
-    node_ids = torch.LongTensor([0, 1, 0])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
+    node_timestamps = torch.IntTensor([6, 7, 8])
+    node_ids = torch.IntTensor([0, 1, 0])
 
     with pytest.raises(TypeError):
         _ = DGData.from_raw(
@@ -272,11 +272,11 @@ def test_init_dg_data_bad_args_non_integral_types():
 
 
 def test_init_dg_data_bad_args_bad_dynamic_node_feats():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     edge_feats = torch.rand(2, 5)
-    node_ids = torch.LongTensor([1, 2, 3])
-    node_timestamps = torch.LongTensor([6, 7, 8])
+    node_ids = torch.IntTensor([1, 2, 3])
+    node_timestamps = torch.IntTensor([6, 7, 8])
 
     with pytest.raises(TypeError):
         _ = DGData.from_raw(
@@ -300,10 +300,10 @@ def test_init_dg_data_bad_args_bad_dynamic_node_feats():
 
 def test_init_dg_data_bad_args_bad_static_node_feats():
     # Num nodes = 21
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
-    node_ids = torch.LongTensor([1, 2, 3])
-    node_timestamps = torch.LongTensor([6, 7, 8])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
+    node_ids = torch.IntTensor([1, 2, 3])
+    node_timestamps = torch.IntTensor([6, 7, 8])
 
     with pytest.raises(TypeError):
         _ = DGData.from_raw(
@@ -322,8 +322,8 @@ def test_init_dg_data_bad_args_bad_static_node_feats():
         )
 
     # Num nodes = 21
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
 
     with pytest.raises(ValueError):
         _ = DGData.from_raw(
@@ -337,10 +337,10 @@ def test_init_dg_data_bad_args_bad_static_node_feats():
         )
 
     # Num nodes = 101
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
-    node_ids = torch.LongTensor([1, 2, 100])
-    node_timestamps = torch.LongTensor([6, 7, 8])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
+    node_ids = torch.IntTensor([1, 2, 100])
+    node_timestamps = torch.IntTensor([6, 7, 8])
 
     with pytest.raises(ValueError):
         _ = DGData.from_raw(
@@ -355,8 +355,8 @@ def test_init_dg_data_bad_args_bad_static_node_feats():
 
 
 def test_from_csv_with_edge_features():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    timestamps = torch.IntTensor([1, 5])
     edge_feats = torch.rand(2, 5)
     data = DGData.from_raw(
         edge_timestamps=timestamps, edge_index=edge_index, edge_feats=edge_feats
@@ -402,10 +402,10 @@ def test_from_csv_with_edge_features():
 
 
 def test_from_csv_with_node_events():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 1])
-    node_ids = torch.LongTensor([7, 8])
-    node_timestamps = torch.LongTensor([3, 6])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 1])
+    node_ids = torch.IntTensor([7, 8])
+    node_timestamps = torch.IntTensor([3, 6])
     data = DGData.from_raw(
         edge_timestamps=edge_timestamps,
         edge_index=edge_index,
@@ -474,10 +474,10 @@ def test_from_csv_with_node_events():
 
 
 def test_from_csv_with_node_features():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 1])
-    node_ids = torch.LongTensor([7, 8])
-    node_timestamps = torch.LongTensor([3, 6])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 1])
+    node_ids = torch.IntTensor([7, 8])
+    node_timestamps = torch.IntTensor([3, 6])
     dynamic_node_feats = torch.rand(2, 5)
     static_node_feats = torch.rand(21, 3)
 
@@ -806,8 +806,8 @@ def test_from_tgbn(mock_dataset_cls, tgb_dataset_factory):
 
 
 def test_discretize_reduce_op_first():
-    edge_index = torch.LongTensor([[1, 2], [1, 2], [2, 3], [1, 2], [4, 5]])
-    edge_timestamps = torch.LongTensor([1, 2, 3, 63, 65])
+    edge_index = torch.IntTensor([[1, 2], [1, 2], [2, 3], [1, 2], [4, 5]])
+    edge_timestamps = torch.IntTensor([1, 2, 3, 63, 65])
     edge_feats = torch.rand(5, 5)
     static_node_feats = torch.rand(6, 11)
     data = DGData.from_raw(
@@ -824,9 +824,9 @@ def test_discretize_reduce_op_first():
     assert data.time_delta == TimeDeltaDG('m')
     assert id(coarse_data) != id(data)
 
-    exp_timestamps = torch.LongTensor([0, 0, 1, 1])
-    exp_edge_event_idx = torch.LongTensor([0, 1, 2, 3])
-    exp_edge_index = torch.LongTensor([[1, 2], [2, 3], [1, 2], [4, 5]])
+    exp_timestamps = torch.IntTensor([0, 0, 1, 1])
+    exp_edge_event_idx = torch.IntTensor([0, 1, 2, 3])
+    exp_edge_index = torch.IntTensor([[1, 2], [2, 3], [1, 2], [4, 5]])
     exp_edge_feats = torch.stack(
         [edge_feats[0], edge_feats[2], edge_feats[3], edge_feats[4]]
     )
@@ -844,12 +844,12 @@ def test_discretize_reduce_op_first():
 
 
 def test_discretize_with_node_events_reduce_op_first():
-    edge_index = torch.LongTensor([[1, 2], [1, 2], [2, 3], [1, 2], [4, 5]])
-    edge_timestamps = torch.LongTensor([1, 2, 3, 63, 65])
+    edge_index = torch.IntTensor([[1, 2], [1, 2], [2, 3], [1, 2], [4, 5]])
+    edge_timestamps = torch.IntTensor([1, 2, 3, 63, 65])
     edge_feats = torch.rand(5, 5)
 
-    node_ids = torch.LongTensor([6, 6, 7, 6, 6, 7])
-    node_timestamps = torch.LongTensor([10, 20, 30, 70, 80, 90])
+    node_ids = torch.IntTensor([6, 6, 7, 6, 6, 7])
+    node_timestamps = torch.IntTensor([10, 20, 30, 70, 80, 90])
     dynamic_node_feats = torch.rand(6, 5)
     static_node_feats = torch.rand(8, 11)
     data = DGData.from_raw(
@@ -870,16 +870,16 @@ def test_discretize_with_node_events_reduce_op_first():
     assert data.time_delta == TimeDeltaDG('m')
     assert id(coarse_data) != id(data)
 
-    exp_timestamps = torch.LongTensor([0, 0, 0, 0, 1, 1, 1, 1])
-    exp_edge_event_idx = torch.LongTensor([0, 1, 4, 5])
-    exp_edge_index = torch.LongTensor([[1, 2], [2, 3], [1, 2], [4, 5]])
+    exp_timestamps = torch.IntTensor([0, 0, 0, 0, 1, 1, 1, 1])
+    exp_edge_event_idx = torch.IntTensor([0, 1, 4, 5])
+    exp_edge_index = torch.IntTensor([[1, 2], [2, 3], [1, 2], [4, 5]])
     exp_edge_feats = torch.stack(
         [edge_feats[0], edge_feats[2], edge_feats[3], edge_feats[4]]
     )
     exp_static_node_feats = static_node_feats
 
-    exp_node_event_idx = torch.LongTensor([2, 3, 6, 7])
-    exp_node_ids = torch.LongTensor([6, 7, 6, 7])
+    exp_node_event_idx = torch.IntTensor([2, 3, 6, 7])
+    exp_node_ids = torch.IntTensor([6, 7, 6, 7])
 
     exp_dynamic_node_feats = torch.stack(
         [
@@ -902,8 +902,8 @@ def test_discretize_with_node_events_reduce_op_first():
 
 
 def test_discretize_no_op():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
     data = DGData.from_raw(edge_timestamps, edge_index)
     coarse_data = data.discretize('r')
     assert id(data) != id(coarse_data)  # No Shared memory
@@ -914,8 +914,8 @@ def test_discretize_no_op():
 
 
 def test_discretize_bad_args():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
 
     with pytest.raises(EventOrderedConversionError):
         data = DGData.from_raw(edge_timestamps, edge_index, time_delta='r')
@@ -984,8 +984,8 @@ def test_split_cannot_override_tgb_split():
 
 
 def test_clone():
-    edge_index = torch.LongTensor([[2, 3], [10, 20]])
-    edge_timestamps = torch.LongTensor([1, 5])
+    edge_index = torch.IntTensor([[2, 3], [10, 20]])
+    edge_timestamps = torch.IntTensor([1, 5])
 
     dg1 = DGData.from_raw(edge_timestamps, edge_index)
     dg2 = dg1.clone()
